Initial parameter library integration with main avt_341 package. Also added topic remapping and environment variable yaml config. Added parameter reference and python templating in input yaml files.

+++ avt_341/include/avt_341/mission/formation_path_generator.h
@@ -31,7 +31,7 @@ class FormationPathGenerator{
 													avt_341::msg::FollowerStatus status, Vec2d leaderVx, Vec2d leaderVy);
 
 	// control parameters
-	float gpp2_; // square of global_path_points_dist_
+	double gpp2_; // square of global_path_points_dist_
 
 	// outputs / messages published
 	avt_341::msg::Path desired_global_path_;
